+++ sdk/resourcemanager/servicefabricmanagedclusters/armservicefabricmanagedclusters/applications_client.go
@@ -42,7 +42,7 @@ func NewApplicationsClient(subscriptionID string, credential azcore.TokenCredent
 // BeginCreateOrUpdate - Create or update a Service Fabric managed application resource with the specified name.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-10-01-preview
+// Generated from API version 2026-02-01
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - clusterName - The name of the cluster resource.
 //   - applicationName - The name of the application resource.
@@ -69,7 +69,7 @@ func (client *ApplicationsClient) BeginCreateOrUpdate(ctx context.Context, resou
 // CreateOrUpdate - Create or update a Service Fabric managed application resource with the specified name.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-10-01-preview
+// Generated from API version 2026-02-01
 func (client *ApplicationsClient) createOrUpdate(ctx context.Context, resourceGroupName string, clusterName string, applicationName string, parameters ApplicationResource, options *ApplicationsClientBeginCreateOrUpdateOptions) (*http.Response, error) {
 	var err error
 	const operationName = "ApplicationsClient.BeginCreateOrUpdate"
@@ -115,7 +115,7 @@ func (client *ApplicationsClient) createOrUpdateCreateRequest(ctx context.Contex
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-10-01-preview")
+	reqQP.Set("api-version", "2026-02-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	req.Raw().Header["Content-Type"] = []string{"application/json"}
@@ -128,7 +128,7 @@ func (client *ApplicationsClient) createOrUpdateCreateRequest(ctx context.Contex
 // BeginDelete - Delete a Service Fabric managed application resource with the specified name.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-10-01-preview
+// Generated from API version 2026-02-01
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - clusterName - The name of the cluster resource.
 //   - applicationName - The name of the application resource.
@@ -154,7 +154,7 @@ func (client *ApplicationsClient) BeginDelete(ctx context.Context, resourceGroup
 // Delete - Delete a Service Fabric managed application resource with the specified name.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-10-01-preview
+// Generated from API version 2026-02-01
 func (client *ApplicationsClient) deleteOperation(ctx context.Context, resourceGroupName string, clusterName string, applicationName string, options *ApplicationsClientBeginDeleteOptions) (*http.Response, error) {
 	var err error
 	const operationName = "ApplicationsClient.BeginDelete"
@@ -200,7 +200,7 @@ func (client *ApplicationsClient) deleteCreateRequest(ctx context.Context, resou
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-10-01-preview")
+	reqQP.Set("api-version", "2026-02-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	return req, nil
 }
@@ -209,7 +209,7 @@ func (client *ApplicationsClient) deleteCreateRequest(ctx context.Context, resou
 // deployed application.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-10-01-preview
+// Generated from API version 2026-02-01
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - clusterName - The name of the cluster resource.
 //   - applicationName - The name of the application resource.
@@ -237,7 +237,7 @@ func (client *ApplicationsClient) BeginFetchHealth(ctx context.Context, resource
 // application.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-10-01-preview
+// Generated from API version 2026-02-01
 func (client *ApplicationsClient) fetchHealth(ctx context.Context, resourceGroupName string, clusterName string, applicationName string, parameters ApplicationFetchHealthRequest, options *ApplicationsClientBeginFetchHealthOptions) (*http.Response, error) {
 	var err error
 	const operationName = "ApplicationsClient.BeginFetchHealth"
@@ -283,7 +283,7 @@ func (client *ApplicationsClient) fetchHealthCreateRequest(ctx context.Context, 
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-10-01-preview")
+	reqQP.Set("api-version", "2026-02-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Content-Type"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, parameters); err != nil {
@@ -296,7 +296,7 @@ func (client *ApplicationsClient) fetchHealthCreateRequest(ctx context.Context, 
 // cluster resource.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-10-01-preview
+// Generated from API version 2026-02-01
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - clusterName - The name of the cluster resource.
 //   - applicationName - The name of the application resource.
@@ -347,7 +347,7 @@ func (client *ApplicationsClient) getCreateRequest(ctx context.Context, resource
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-10-01-preview")
+	reqQP.Set("api-version", "2026-02-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -365,7 +365,7 @@ func (client *ApplicationsClient) getHandleResponse(resp *http.Response) (Applic
 // NewListPager - Gets all managed application resources created or in the process of being created in the Service Fabric
 // cluster resource.
 //
-// Generated from API version 2025-10-01-preview
+// Generated from API version 2026-02-01
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - clusterName - The name of the cluster resource.
 //   - options - ApplicationsClientListOptions contains the optional parameters for the ApplicationsClient.NewListPager method.
@@ -412,7 +412,7 @@ func (client *ApplicationsClient) listCreateRequest(ctx context.Context, resourc
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-10-01-preview")
+	reqQP.Set("api-version", "2026-02-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -431,7 +431,7 @@ func (client *ApplicationsClient) listHandleResponse(resp *http.Response) (Appli
 // latest application upgrade.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-10-01-preview
+// Generated from API version 2026-02-01
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - clusterName - The name of the cluster resource.
 //   - applicationName - The name of the application resource.
@@ -458,7 +458,7 @@ func (client *ApplicationsClient) BeginReadUpgrade(ctx context.Context, resource
 // application upgrade.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-10-01-preview
+// Generated from API version 2026-02-01
 func (client *ApplicationsClient) readUpgrade(ctx context.Context, resourceGroupName string, clusterName string, applicationName string, options *ApplicationsClientBeginReadUpgradeOptions) (*http.Response, error) {
 	var err error
 	const operationName = "ApplicationsClient.BeginReadUpgrade"
@@ -504,7 +504,7 @@ func (client *ApplicationsClient) readUpgradeCreateRequest(ctx context.Context, 
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-10-01-preview")
+	reqQP.Set("api-version", "2026-02-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	return req, nil
 }
@@ -513,7 +513,7 @@ func (client *ApplicationsClient) readUpgradeCreateRequest(ctx context.Context, 
 // destabilizing operation that should be used with immense care.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-10-01-preview
+// Generated from API version 2026-02-01
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - clusterName - The name of the cluster resource.
 //   - applicationName - The name of the application resource.
@@ -541,7 +541,7 @@ func (client *ApplicationsClient) BeginRestartDeployedCodePackage(ctx context.Co
 // operation that should be used with immense care.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-10-01-preview
+// Generated from API version 2026-02-01
 func (client *ApplicationsClient) restartDeployedCodePackage(ctx context.Context, resourceGroupName string, clusterName string, applicationName string, parameters RestartDeployedCodePackageRequest, options *ApplicationsClientBeginRestartDeployedCodePackageOptions) (*http.Response, error) {
 	var err error
 	const operationName = "ApplicationsClient.BeginRestartDeployedCodePackage"
@@ -587,7 +587,7 @@ func (client *ApplicationsClient) restartDeployedCodePackageCreateRequest(ctx co
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-10-01-preview")
+	reqQP.Set("api-version", "2026-02-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Content-Type"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, parameters); err != nil {
@@ -600,7 +600,7 @@ func (client *ApplicationsClient) restartDeployedCodePackageCreateRequest(ctx co
 // from where it was paused.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-10-01-preview
+// Generated from API version 2026-02-01
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - clusterName - The name of the cluster resource.
 //   - applicationName - The name of the application resource.
@@ -628,7 +628,7 @@ func (client *ApplicationsClient) BeginResumeUpgrade(ctx context.Context, resour
 // where it was paused.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-10-01-preview
+// Generated from API version 2026-02-01
 func (client *ApplicationsClient) resumeUpgrade(ctx context.Context, resourceGroupName string, clusterName string, applicationName string, parameters RuntimeResumeApplicationUpgradeParameters, options *ApplicationsClientBeginResumeUpgradeOptions) (*http.Response, error) {
 	var err error
 	const operationName = "ApplicationsClient.BeginResumeUpgrade"
@@ -674,7 +674,7 @@ func (client *ApplicationsClient) resumeUpgradeCreateRequest(ctx context.Context
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-10-01-preview")
+	reqQP.Set("api-version", "2026-02-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Content-Type"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, parameters); err != nil {
@@ -687,7 +687,7 @@ func (client *ApplicationsClient) resumeUpgradeCreateRequest(ctx context.Context
 // the application to the previous version.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-10-01-preview
+// Generated from API version 2026-02-01
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - clusterName - The name of the cluster resource.
 //   - applicationName - The name of the application resource.
@@ -714,7 +714,7 @@ func (client *ApplicationsClient) BeginStartRollback(ctx context.Context, resour
 // application to the previous version.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-10-01-preview
+// Generated from API version 2026-02-01
 func (client *ApplicationsClient) startRollback(ctx context.Context, resourceGroupName string, clusterName string, applicationName string, options *ApplicationsClientBeginStartRollbackOptions) (*http.Response, error) {
 	var err error
 	const operationName = "ApplicationsClient.BeginStartRollback"
@@ -760,7 +760,7 @@ func (client *ApplicationsClient) startRollbackCreateRequest(ctx context.Context
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-10-01-preview")
+	reqQP.Set("api-version", "2026-02-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	return req, nil
 }
@@ -768,7 +768,7 @@ func (client *ApplicationsClient) startRollbackCreateRequest(ctx context.Context
 // BeginUpdate - Updates an application resource of a given managed cluster.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-10-01-preview
+// Generated from API version 2026-02-01
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - clusterName - The name of the cluster resource.
 //   - applicationName - The name of the application resource.
@@ -795,7 +795,7 @@ func (client *ApplicationsClient) BeginUpdate(ctx context.Context, resourceGroup
 // Update - Updates an application resource of a given managed cluster.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-10-01-preview
+// Generated from API version 2026-02-01
 func (client *ApplicationsClient) update(ctx context.Context, resourceGroupName string, clusterName string, applicationName string, parameters ApplicationUpdateParameters, options *ApplicationsClientBeginUpdateOptions) (*http.Response, error) {
 	var err error
 	const operationName = "ApplicationsClient.BeginUpdate"
@@ -841,7 +841,7 @@ func (client *ApplicationsClient) updateCreateRequest(ctx context.Context, resou
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-10-01-preview")
+	reqQP.Set("api-version", "2026-02-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	req.Raw().Header["Content-Type"] = []string{"application/json"}
@@ -854,7 +854,7 @@ func (client *ApplicationsClient) updateCreateRequest(ctx context.Context, resou
 // BeginUpdateUpgrade - Send a request to update the current application upgrade.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-10-01-preview
+// Generated from API version 2026-02-01
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - clusterName - The name of the cluster resource.
 //   - applicationName - The name of the application resource.
@@ -881,7 +881,7 @@ func (client *ApplicationsClient) BeginUpdateUpgrade(ctx context.Context, resour
 // UpdateUpgrade - Send a request to update the current application upgrade.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-10-01-preview
+// Generated from API version 2026-02-01
 func (client *ApplicationsClient) updateUpgrade(ctx context.Context, resourceGroupName string, clusterName string, applicationName string, parameters RuntimeUpdateApplicationUpgradeParameters, options *ApplicationsClientBeginUpdateUpgradeOptions) (*http.Response, error) {
 	var err error
 	const operationName = "ApplicationsClient.BeginUpdateUpgrade"
@@ -927,7 +927,7 @@ func (client *ApplicationsClient) updateUpgradeCreateRequest(ctx context.Context
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-10-01-preview")
+	reqQP.Set("api-version", "2026-02-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Content-Type"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, parameters); err != nil {