+++ sdk/resourcemanager/servicefabricmanagedclusters/armservicefabricmanagedclusters/models_serde.go
@@ -1162,235 +1162,6 @@ func (e *ErrorModelError) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
-// MarshalJSON implements the json.Marshaller interface for type FaultSimulation.
-func (f FaultSimulation) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]any)
-	populate(objectMap, "details", f.Details)
-	populateTime[datetime.RFC3339](objectMap, "endTime", f.EndTime)
-	populate(objectMap, "simulationId", f.SimulationID)
-	populateTime[datetime.RFC3339](objectMap, "startTime", f.StartTime)
-	populate(objectMap, "status", f.Status)
-	return json.Marshal(objectMap)
-}
-
-// UnmarshalJSON implements the json.Unmarshaller interface for type FaultSimulation.
-func (f *FaultSimulation) UnmarshalJSON(data []byte) error {
-	var rawMsg map[string]json.RawMessage
-	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return fmt.Errorf("unmarshalling type %T: %v", f, err)
-	}
-	for key, val := range rawMsg {
-		var err error
-		switch key {
-		case "details":
-			err = unpopulate(val, "Details", &f.Details)
-			delete(rawMsg, key)
-		case "endTime":
-			err = unpopulateTime[datetime.RFC3339](val, "EndTime", &f.EndTime)
-			delete(rawMsg, key)
-		case "simulationId":
-			err = unpopulate(val, "SimulationID", &f.SimulationID)
-			delete(rawMsg, key)
-		case "startTime":
-			err = unpopulateTime[datetime.RFC3339](val, "StartTime", &f.StartTime)
-			delete(rawMsg, key)
-		case "status":
-			err = unpopulate(val, "Status", &f.Status)
-			delete(rawMsg, key)
-		}
-		if err != nil {
-			return fmt.Errorf("unmarshalling type %T: %v", f, err)
-		}
-	}
-	return nil
-}
-
-// MarshalJSON implements the json.Marshaller interface for type FaultSimulationConstraints.
-func (f FaultSimulationConstraints) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]any)
-	populateTime[datetime.RFC3339](objectMap, "expirationTime", f.ExpirationTime)
-	return json.Marshal(objectMap)
-}
-
-// UnmarshalJSON implements the json.Unmarshaller interface for type FaultSimulationConstraints.
-func (f *FaultSimulationConstraints) UnmarshalJSON(data []byte) error {
-	var rawMsg map[string]json.RawMessage
-	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return fmt.Errorf("unmarshalling type %T: %v", f, err)
-	}
-	for key, val := range rawMsg {
-		var err error
-		switch key {
-		case "expirationTime":
-			err = unpopulateTime[datetime.RFC3339](val, "ExpirationTime", &f.ExpirationTime)
-			delete(rawMsg, key)
-		}
-		if err != nil {
-			return fmt.Errorf("unmarshalling type %T: %v", f, err)
-		}
-	}
-	return nil
-}
-
-// MarshalJSON implements the json.Marshaller interface for type FaultSimulationContent.
-func (f FaultSimulationContent) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]any)
-	populate(objectMap, "constraints", f.Constraints)
-	populate(objectMap, "faultKind", f.FaultKind)
-	populate(objectMap, "force", f.Force)
-	return json.Marshal(objectMap)
-}
-
-// UnmarshalJSON implements the json.Unmarshaller interface for type FaultSimulationContent.
-func (f *FaultSimulationContent) UnmarshalJSON(data []byte) error {
-	var rawMsg map[string]json.RawMessage
-	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return fmt.Errorf("unmarshalling type %T: %v", f, err)
-	}
-	for key, val := range rawMsg {
-		var err error
-		switch key {
-		case "constraints":
-			err = unpopulate(val, "Constraints", &f.Constraints)
-			delete(rawMsg, key)
-		case "faultKind":
-			err = unpopulate(val, "FaultKind", &f.FaultKind)
-			delete(rawMsg, key)
-		case "force":
-			err = unpopulate(val, "Force", &f.Force)
-			delete(rawMsg, key)
-		}
-		if err != nil {
-			return fmt.Errorf("unmarshalling type %T: %v", f, err)
-		}
-	}
-	return nil
-}
-
-// MarshalJSON implements the json.Marshaller interface for type FaultSimulationContentWrapper.
-func (f FaultSimulationContentWrapper) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]any)
-	populate(objectMap, "parameters", f.Parameters)
-	return json.Marshal(objectMap)
-}
-
-// UnmarshalJSON implements the json.Unmarshaller interface for type FaultSimulationContentWrapper.
-func (f *FaultSimulationContentWrapper) UnmarshalJSON(data []byte) error {
-	var rawMsg map[string]json.RawMessage
-	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return fmt.Errorf("unmarshalling type %T: %v", f, err)
-	}
-	for key, val := range rawMsg {
-		var err error
-		switch key {
-		case "parameters":
-			f.Parameters, err = unmarshalFaultSimulationContentClassification(val)
-			delete(rawMsg, key)
-		}
-		if err != nil {
-			return fmt.Errorf("unmarshalling type %T: %v", f, err)
-		}
-	}
-	return nil
-}
-
-// MarshalJSON implements the json.Marshaller interface for type FaultSimulationDetails.
-func (f FaultSimulationDetails) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]any)
-	populate(objectMap, "clusterId", f.ClusterID)
-	populate(objectMap, "nodeTypeFaultSimulation", f.NodeTypeFaultSimulation)
-	populate(objectMap, "operationId", f.OperationID)
-	populate(objectMap, "parameters", f.Parameters)
-	return json.Marshal(objectMap)
-}
-
-// UnmarshalJSON implements the json.Unmarshaller interface for type FaultSimulationDetails.
-func (f *FaultSimulationDetails) UnmarshalJSON(data []byte) error {
-	var rawMsg map[string]json.RawMessage
-	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return fmt.Errorf("unmarshalling type %T: %v", f, err)
-	}
-	for key, val := range rawMsg {
-		var err error
-		switch key {
-		case "clusterId":
-			err = unpopulate(val, "ClusterID", &f.ClusterID)
-			delete(rawMsg, key)
-		case "nodeTypeFaultSimulation":
-			err = unpopulate(val, "NodeTypeFaultSimulation", &f.NodeTypeFaultSimulation)
-			delete(rawMsg, key)
-		case "operationId":
-			err = unpopulate(val, "OperationID", &f.OperationID)
-			delete(rawMsg, key)
-		case "parameters":
-			f.Parameters, err = unmarshalFaultSimulationContentClassification(val)
-			delete(rawMsg, key)
-		}
-		if err != nil {
-			return fmt.Errorf("unmarshalling type %T: %v", f, err)
-		}
-	}
-	return nil
-}
-
-// MarshalJSON implements the json.Marshaller interface for type FaultSimulationIDContent.
-func (f FaultSimulationIDContent) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]any)
-	populate(objectMap, "simulationId", f.SimulationID)
-	return json.Marshal(objectMap)
-}
-
-// UnmarshalJSON implements the json.Unmarshaller interface for type FaultSimulationIDContent.
-func (f *FaultSimulationIDContent) UnmarshalJSON(data []byte) error {
-	var rawMsg map[string]json.RawMessage
-	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return fmt.Errorf("unmarshalling type %T: %v", f, err)
-	}
-	for key, val := range rawMsg {
-		var err error
-		switch key {
-		case "simulationId":
-			err = unpopulate(val, "SimulationID", &f.SimulationID)
-			delete(rawMsg, key)
-		}
-		if err != nil {
-			return fmt.Errorf("unmarshalling type %T: %v", f, err)
-		}
-	}
-	return nil
-}
-
-// MarshalJSON implements the json.Marshaller interface for type FaultSimulationListResult.
-func (f FaultSimulationListResult) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]any)
-	populate(objectMap, "nextLink", f.NextLink)
-	populate(objectMap, "value", f.Value)
-	return json.Marshal(objectMap)
-}
-
-// UnmarshalJSON implements the json.Unmarshaller interface for type FaultSimulationListResult.
-func (f *FaultSimulationListResult) UnmarshalJSON(data []byte) error {
-	var rawMsg map[string]json.RawMessage
-	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return fmt.Errorf("unmarshalling type %T: %v", f, err)
-	}
-	for key, val := range rawMsg {
-		var err error
-		switch key {
-		case "nextLink":
-			err = unpopulate(val, "NextLink", &f.NextLink)
-			delete(rawMsg, key)
-		case "value":
-			err = unpopulate(val, "Value", &f.Value)
-			delete(rawMsg, key)
-		}
-		if err != nil {
-			return fmt.Errorf("unmarshalling type %T: %v", f, err)
-		}
-	}
-	return nil
-}
-
 // MarshalJSON implements the json.Marshaller interface for type FrontendConfiguration.
 func (f FrontendConfiguration) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
@@ -1845,6 +1616,7 @@ func (m ManagedClusterProperties) MarshalJSON() ([]byte, error) {
 	populate(objectMap, "publicIPPrefixId", m.PublicIPPrefixID)
 	populate(objectMap, "publicIPv6PrefixId", m.PublicIPv6PrefixID)
 	populate(objectMap, "serviceEndpoints", m.ServiceEndpoints)
+	populate(objectMap, "skipManagedNsgAssignment", m.SkipManagedNsgAssignment)
 	populate(objectMap, "subnetId", m.SubnetID)
 	populate(objectMap, "upgradeDescription", m.UpgradeDescription)
 	populate(objectMap, "useCustomVnet", m.UseCustomVnet)
@@ -1973,6 +1745,9 @@ func (m *ManagedClusterProperties) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "serviceEndpoints":
 			err = unpopulate(val, "ServiceEndpoints", &m.ServiceEndpoints)
+			delete(rawMsg, key)
+		case "skipManagedNsgAssignment":
+			err = unpopulate(val, "SkipManagedNsgAssignment", &m.SkipManagedNsgAssignment)
 			delete(rawMsg, key)
 		case "subnetId":
 			err = unpopulate(val, "SubnetID", &m.SubnetID)
@@ -2453,45 +2228,6 @@ func (n *NodeTypeAvailableSKU) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
-// MarshalJSON implements the json.Marshaller interface for type NodeTypeFaultSimulation.
-func (n NodeTypeFaultSimulation) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]any)
-	populate(objectMap, "nodeTypeName", n.NodeTypeName)
-	populate(objectMap, "operationId", n.OperationID)
-	populate(objectMap, "operationStatus", n.OperationStatus)
-	populate(objectMap, "status", n.Status)
-	return json.Marshal(objectMap)
-}
-
-// UnmarshalJSON implements the json.Unmarshaller interface for type NodeTypeFaultSimulation.
-func (n *NodeTypeFaultSimulation) UnmarshalJSON(data []byte) error {
-	var rawMsg map[string]json.RawMessage
-	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return fmt.Errorf("unmarshalling type %T: %v", n, err)
-	}
-	for key, val := range rawMsg {
-		var err error
-		switch key {
-		case "nodeTypeName":
-			err = unpopulate(val, "NodeTypeName", &n.NodeTypeName)
-			delete(rawMsg, key)
-		case "operationId":
-			err = unpopulate(val, "OperationID", &n.OperationID)
-			delete(rawMsg, key)
-		case "operationStatus":
-			err = unpopulate(val, "OperationStatus", &n.OperationStatus)
-			delete(rawMsg, key)
-		case "status":
-			err = unpopulate(val, "Status", &n.Status)
-			delete(rawMsg, key)
-		}
-		if err != nil {
-			return fmt.Errorf("unmarshalling type %T: %v", n, err)
-		}
-	}
-	return nil
-}
-
 // MarshalJSON implements the json.Marshaller interface for type NodeTypeListResult.
 func (n NodeTypeListResult) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
@@ -2606,6 +2342,7 @@ func (n NodeTypeProperties) MarshalJSON() ([]byte, error) {
 	populate(objectMap, "enableNodePublicIP", n.EnableNodePublicIP)
 	populate(objectMap, "enableNodePublicIPv6", n.EnableNodePublicIPv6)
 	populate(objectMap, "enableOverProvisioning", n.EnableOverProvisioning)
+	populate(objectMap, "enableResilientEphemeralOsDisk", n.EnableResilientEphemeralOsDisk)
 	populate(objectMap, "ephemeralPorts", n.EphemeralPorts)
 	populate(objectMap, "evictionPolicy", n.EvictionPolicy)
 	populate(objectMap, "frontendConfigurations", n.FrontendConfigurations)
@@ -2698,6 +2435,9 @@ func (n *NodeTypeProperties) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "enableOverProvisioning":
 			err = unpopulate(val, "EnableOverProvisioning", &n.EnableOverProvisioning)
+			delete(rawMsg, key)
+		case "enableResilientEphemeralOsDisk":
+			err = unpopulate(val, "EnableResilientEphemeralOsDisk", &n.EnableResilientEphemeralOsDisk)
 			delete(rawMsg, key)
 		case "ephemeralPorts":
 			err = unpopulate(val, "EphemeralPorts", &n.EphemeralPorts)
@@ -4881,45 +4621,6 @@ func (v *VmssDataDisk) UnmarshalJSON(data []byte) error {
 		}
 		if err != nil {
 			return fmt.Errorf("unmarshalling type %T: %v", v, err)
-		}
-	}
-	return nil
-}
-
-// MarshalJSON implements the json.Marshaller interface for type ZoneFaultSimulationContent.
-func (z ZoneFaultSimulationContent) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]any)
-	populate(objectMap, "constraints", z.Constraints)
-	objectMap["faultKind"] = FaultKindZone
-	populate(objectMap, "force", z.Force)
-	populate(objectMap, "zones", z.Zones)
-	return json.Marshal(objectMap)
-}
-
-// UnmarshalJSON implements the json.Unmarshaller interface for type ZoneFaultSimulationContent.
-func (z *ZoneFaultSimulationContent) UnmarshalJSON(data []byte) error {
-	var rawMsg map[string]json.RawMessage
-	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return fmt.Errorf("unmarshalling type %T: %v", z, err)
-	}
-	for key, val := range rawMsg {
-		var err error
-		switch key {
-		case "constraints":
-			err = unpopulate(val, "Constraints", &z.Constraints)
-			delete(rawMsg, key)
-		case "faultKind":
-			err = unpopulate(val, "FaultKind", &z.FaultKind)
-			delete(rawMsg, key)
-		case "force":
-			err = unpopulate(val, "Force", &z.Force)
-			delete(rawMsg, key)
-		case "zones":
-			err = unpopulate(val, "Zones", &z.Zones)
-			delete(rawMsg, key)
-		}
-		if err != nil {
-			return fmt.Errorf("unmarshalling type %T: %v", z, err)
 		}
 	}
 	return nil